requirements.txt
boto3==1.16.6
botocore==1.19.6
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==7.2.0
pycryptodome==3.9.9
python-dateutil==2.8.1
python-gnupg==0.4.6
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
